+++ src/plugins/lyrics-genius/renderer.ts
@@ -38,7 +38,7 @@ export const onRendererLoad = ({
 
   let unregister: (() => void) | null = null;
 
-  on('ytmd:update-song-info', (extractedSongInfo: SongInfo) => {
+  on('ytd:update-song-info', (extractedSongInfo: SongInfo) => {
     unregister?.();
 
     setTimeout(async () => {