+++ src/reset.d.ts
@@ -15,7 +15,7 @@ declare global {
   }
 
   interface DocumentEventMap {
-    'ytmd:audio-can-play': CustomEvent<Compressor>;
+    'ytd:audio-can-play': CustomEvent<Compressor>;
     videodatachange: CustomEvent<VideoDataChanged>;
   }
 
@@ -26,7 +26,7 @@ declare global {
     electronIs: typeof is;
     ELECTRON_RENDERER_URL: string | undefined;
     /**
-     * YouTube Music internal variable (Last interaction time)
+     * YouTube internal variable (Last interaction time)
      */
     _lact: number;
     navigation: Navigation;